+++ chat/app/ui/profile_wizzard.py
@@ -427,12 +427,12 @@ class ProfileWizard(BasePage):
                         "personality": personality
                     }
                 }
-                Backend.get_instance().generate(prompt)
+                Backend.generate(prompt)
             else:
                 data = {
                     "profile": json.dumps(profile)
                 }
-                Backend.get_instance().update_user_data(data)
+                Backend.update_user_data(data)
                 self.navigator("contacts")
 
     def prev_page(self):
@@ -464,8 +464,8 @@ class ProfileWizard(BasePage):
         self.clear_pages()
 
         if self.user_profile_mode:
-            data = Backend.get_instance().get_user_data()
-            name = Backend.get_instance().get_user_name()
+            data = Backend.get_user_data()
+            name = Backend.get_user_name()
             if data and "profile" in data:
                 profile = data["profile"]
             else:
@@ -485,11 +485,11 @@ class ProfileWizard(BasePage):
             #self.add_page(self.progress_page)
 
         if not self.user_profile_mode:
-            Backend.get_instance().register_incomming_message(self.on_incomming_message)
+            Backend.register_incomming_message(self.on_incomming_message)
 
     def on_leave(self):
         if not self.user_profile_mode:
-            Backend.get_instance().unregister_incomming_message(self.on_incomming_message)
+            Backend.unregister_incomming_message(self.on_incomming_message)
 
     def on_incomming_message(self, msg: dict):
         try: